+++ packages/atomic/cypress/integration/facets/facet/facet.cypress.ts
@@ -1073,15 +1073,38 @@ describe('Facet v1 Test Suites', () => {
   });
 
   describe('with allowed-values', () => {
-    it('returns only allowed values', () => {
+    beforeEach(() => {
       new TestFixture()
         .with(addFacet({field: 'objecttype', 'allowed-values': 'FAQ,File'}))
         .init();
+    });
 
+    it('returns only allowed values', () => {
       FacetSelectors.values()
         .should('contain.text', 'FAQ')
         .should('contain.text', 'File')
         .should('not.contain.text', 'Message');
     });
+
+    CommonAssertions.assertConsoleWarningMessage(
+      'This will be enforced in the next major versio'
+    );
+  });
+
+  describe('with allowed-values as JSON string array', () => {
+    beforeEach(() => {
+      new TestFixture()
+        .with(
+          addFacet({field: 'objecttype', 'allowed-values': '["FAQ","File"]'})
+        )
+        .init();
+    });
+    it('returns only allowed values', () => {
+      FacetSelectors.values()
+        .should('contain.text', 'FAQ')
+        .should('contain.text', 'File')
+        .should('not.contain.text', 'Message');
+    });
+    CommonAssertions.assertConsoleWarning(false);
   });
 });