+++ packages/atomic/src/components/insight/atomic-insight-numeric-facet/atomic-insight-numeric-facet.tsx
@@ -60,7 +60,7 @@ export class AtomicInsightNumericFacet
   private manualRanges: NumericRangeWithLabel[] = [];
   @Element() private host!: HTMLElement;
   private formatter: NumberFormatter = defaultNumberFormatter;
-  private numericFacetCommon!: NumericFacetCommon;
+  private numericFacetCommon?: NumericFacetCommon;
   @BindStateToController('facetForRange')
   @State()
   public facetState!: InsightNumericFacetState;
@@ -189,7 +189,7 @@ export class AtomicInsightNumericFacet
   }
 
   public disconnectedCallback() {
-    this.numericFacetCommon.disconnectedCallback();
+    this.numericFacetCommon?.disconnectedCallback();
   }
 
   private initializeFacetForInput() {