+++ packages/atomic/src/components/search/result-lists/atomic-result-children/atomic-result-children.tsx
@@ -1,12 +1,20 @@
 import {
-  FoldedCollection,
   FoldedResult,
   FoldedResultList,
   FoldedResultListState,
   Result,
   ResultTemplatesManager,
 } from '@coveo/headless';
-import {Component, Element, State, h, Listen, Prop, Host} from '@stencil/core';
+import {
+  Component,
+  Element,
+  State,
+  h,
+  Listen,
+  Prop,
+  Host,
+  VNode,
+} from '@stencil/core';
 import {InitializeBindings} from '../../../../utils/initialization-utils';
 import {elementHasAncestorTag} from '../../../../utils/utils';
 import {
@@ -16,8 +24,11 @@ import {
   ResultDisplayConfigContext,
   DisplayConfig,
 } from '../../result-template-components/result-template-decorators';
-import {BaseResultList, ResultListCommon} from '../result-list-common';
-import {TemplateContent} from '../../result-templates/result-template-common';
+import {
+  BaseResultList,
+  ResultListCommon,
+} from '../../../common/result-list/result-list-common';
+import {TemplateContent} from '../../../common/result-templates/result-template-common';
 import {
   FoldedResultListContext,
   FoldedResultListStateContext,
@@ -36,40 +47,36 @@ const componentTag = 'atomic-result-children';
  * only when children exist.
  * @part no-result-root - The wrapper for the message when there are no results.
  * @part show-hide-button - The button that allows to collapse or show all child results.
+ * @slot before-children - Slot that allows rendering content before the list of children, only when children exist.
+ * @slot after-children - Slot that allows rendering content after the list of children, only when children exist.
  */
 @Component({
   tag: 'atomic-result-children',
   styleUrl: 'atomic-result-children.pcss',
   shadow: true,
 })
-export class AtomicResultChildren implements BaseResultList {
+export class AtomicResultChildren implements BaseResultList<Bindings> {
   @InitializeBindings() public bindings!: Bindings;
+  public resultListCommon!: ResultListCommon<Bindings>;
   @ChildTemplatesContext()
   public templatesManager!: ResultTemplatesManager<TemplateContent>;
+  @FoldedResultListContext()
+  private foldedResultList!: FoldedResultList;
   @ResultContext({folded: true})
   private result!: FoldedResult;
   @ResultDisplayConfigContext()
   private displayConfig!: DisplayConfig;
-
-  public resultListCommon!: ResultListCommon;
+  private initialChildren!: FoldedResult[];
 
   @Element() public host!: HTMLDivElement;
   @State() public error!: Error;
   @State() public ready = false;
   @State() public templateHasError = false;
-
-  @FoldedResultListContext()
-  private foldedResultList!: FoldedResultList;
-
   @FoldedResultListStateContext()
   @State()
   private foldedResultListState!: FoldedResultListState;
-
   @State()
-  private hideResults = false;
-
-  @State()
-  private showAllResults = false;
+  private showInitialChildren = false;
 
   /**
    * Whether to inherit templates defined in a parent atomic-result-children. Only works for the second level of child nesting.
@@ -80,13 +87,9 @@ export class AtomicResultChildren implements BaseResultList {
    */
   @Prop({reflect: true}) imageSize?: ResultDisplayImageSize;
   /**
-   * The copy for an empty result state.
-   *
-   * @defaultValue `No documents are related to this one.`
+   * The non-localized copy for an empty result state.
    */
-  @Prop() public noResultText = '';
-
-  private initialChildren!: FoldedResult[];
+  @Prop() public noResultText = 'no-documents-related';
 
   @Listen('atomic/resolveChildTemplates')
   public resolveChildTemplates(event: ChildTemplatesContextEvent) {
@@ -97,9 +100,10 @@ export class AtomicResultChildren implements BaseResultList {
   public async initialize() {
     const childrenTemplates = Array.from(
       this.host.querySelectorAll(childTemplateComponent)
-    ).filter((template) => {
-      return !elementHasAncestorTag(template, childTemplateComponent);
-    });
+    ).filter(
+      (template) => !elementHasAncestorTag(template, childTemplateComponent)
+    );
+
     if (!childrenTemplates.length && !this.inheritTemplates) {
       this.error = new Error(
         `The "${componentTag}" component requires at least one "${childTemplateComponent}" component.`
@@ -137,7 +141,7 @@ export class AtomicResultChildren implements BaseResultList {
     return fragment;
   }
 
-  private getContent(result: Result) {
+  private getTemplatedContent(result: Result) {
     return (
       this.resultListCommon!.resultTemplatesManager.selectTemplate(result) ||
       this.selectInheritedTemplate(result)
@@ -145,7 +149,7 @@ export class AtomicResultChildren implements BaseResultList {
   }
 
   private renderChild(child: FoldedResult, isLast: boolean) {
-    const content = this.getContent(child.result);
+    const content = this.getTemplatedContent(child.result);
     if (!content) {
       return null;
     }
@@ -167,97 +171,73 @@ export class AtomicResultChildren implements BaseResultList {
     );
   }
 
-  private getResult() {
+  private get collection() {
+    return this.foldedResultListState.results.find((r) => {
+      return r.result.uniqueId === this.result.result.uniqueId;
+    });
+  }
+
+  private get foldedResult(): FoldedResult {
+    return this.collection || this.result;
+  }
+
+  private get numberOfChildren() {
+    return this.foldedResult.children.length;
+  }
+
+  private get hasChildren() {
+    return !!this.numberOfChildren;
+  }
+
+  public componentWillRender() {
+    const collection = this.collection;
+    if (this.initialChildren || !collection) {
+      return;
+    }
+
+    this.initialChildren = collection.children;
+  }
+
+  private renderPlaceholder() {
     return (
-      this.foldedResultListState.results.find((r) => {
-        return r.result.uniqueId === this.result.result.uniqueId;
-      }) || (this.result as FoldedCollection)
+      <ListDisplayResultsPlaceholder
+        numberOfPlaceholders={this.numberOfChildren || 2}
+        density={this.displayConfig.density}
+        imageSize={this.imageSize || this.displayConfig.imageSize}
+        isChild
+      />
     );
   }
 
-  private shouldShowChildrenWrapper() {
-    const result = this.getResult();
-    if (!this.initialChildren.length && this.hideResults) {
-      return false;
-    }
+  private renderNoResult() {
     return (
-      Boolean(result.children.length) ||
-      result.isLoadingMoreResults ||
-      this.showAllResults
+      <p part="no-result-root" class="no-result-root">
+        {this.bindings.i18n.t(this.noResultText)}
+      </p>
     );
-  }
-
-  public componentWillUpdate() {
-    if (this.getResult().children && !this.initialChildren) {
-      this.initialChildren = this.getResult().children;
-    }
-    if (this.getResult().isLoadingMoreResults && !this.showAllResults) {
-      this.showAllResults = true;
-    }
-  }
-
-  private get hasResults() {
-    return this.getResult().children.length !== 0;
-  }
-
-  private getComponents() {
-    const result = this.getResult();
-    if (result.isLoadingMoreResults) {
-      return (
-        <ListDisplayResultsPlaceholder
-          resultsPerPageState={{
-            numberOfResults: result.children.length || 2,
-          }}
-          density={this.displayConfig.density}
-          imageSize={this.imageSize || this.displayConfig.imageSize}
-          isChild
-        />
-      );
-    }
-    if (this.hasResults) {
-      const children = this.hideResults
-        ? this.initialChildren
-        : result.children;
-      return children.map((child, i) =>
-        this.renderChild(child, i === children.length - 1)
-      );
-    }
-    if (this.showAllResults) {
-      return (
-        <p part="no-result-root" class="no-result-root">
-          {this.noResultText || this.bindings.i18n.t('no-documents-related')}
-        </p>
-      );
-    }
-    return null;
   }
 
   private renderCollapseButton() {
-    const result = this.getResult();
-    if (
-      Boolean(result.children.length) &&
-      result.children.length !== this.initialChildren.length &&
-      !result.isLoadingMoreResults &&
-      !result.moreResultsAvailable
-    ) {
+    const collection = this.collection!;
+    if (!!collection.children.length && !collection.moreResultsAvailable) {
       return (
         <Button
           part="show-hide-button"
           class="show-hide-button"
           style="text-primary"
           onClick={() => {
-            if (this.hideResults) {
+            if (this.showInitialChildren) {
               this.foldedResultList.logShowMoreFoldedResults(
                 this.result.result
               );
             } else {
               this.foldedResultList.logShowLessFoldedResults();
             }
-            this.hideResults = !this.hideResults;
+            this.showInitialChildren = !this.showInitialChildren;
           }}
         >
           {this.bindings.i18n.t(
-            this.hideResults ? 'load-all-results' : 'collapse-results'
+            this.showInitialChildren ? 'load-all-results' : 'collapse-results'
           )}
         </Button>
       );
@@ -265,24 +245,70 @@ export class AtomicResultChildren implements BaseResultList {
     return null;
   }
 
+  private renderChildrenWrapper(content: VNode | VNode[]) {
+    return (
+      <div part="children-root">
+        {this.hasChildren && <slot name="before-children"></slot>}
+        {content}
+        {this.hasChildren && <slot name="after-children"></slot>}
+      </div>
+    );
+  }
+
+  private renderChildren(children: FoldedResult[]) {
+    return this.renderChildrenWrapper(
+      children.map((child, i) =>
+        this.renderChild(child, i === children.length - 1)
+      )
+    );
+  }
+
+  private renderCollection() {
+    const collection = this.collection!;
+
+    if (collection.isLoadingMoreResults) {
+      return this.renderChildrenWrapper(this.renderPlaceholder());
+    }
+
+    if (!collection.moreResultsAvailable && !this.hasChildren) {
+      return this.renderNoResult();
+    }
+
+    if (!this.hasChildren) {
+      return;
+    }
+
+    const childrenToRender = this.showInitialChildren
+      ? this.initialChildren
+      : collection.children;
+
+    return (
+      <Host>
+        {this.renderCollapseButton()}
+        {this.renderChildren(childrenToRender)}
+      </Host>
+    );
+  }
+
+  private renderFoldedResult() {
+    if (!this.hasChildren) {
+      return;
+    }
+
+    return this.renderChildren(this.foldedResult.children);
+  }
+
   public render() {
     if (!this.ready) {
       return null;
     }
+
     if (this.templateHasError) {
       return <slot></slot>;
     }
-    return (
-      <Host>
-        {this.renderCollapseButton()}
-        {this.shouldShowChildrenWrapper() && (
-          <div part="children-root">
-            {this.hasResults && <slot name="before-children"></slot>}
-            {this.getComponents()}
-            {this.hasResults && <slot name="after-children"></slot>}
-          </div>
-        )}
-      </Host>
-    );
+
+    return this.collection
+      ? this.renderCollection()
+      : this.renderFoldedResult();
   }
 }