+++ packages/atomic/src/pages/header.js
@@ -9,6 +9,10 @@ const links = [
   {href: '/examples/suggestions.html', label: 'Custom Query Suggestions'},
   {href: '/examples/instantResults.html', label: 'Instant Results'},
   {href: '/examples/insights.html', label: 'Insights Panel'},
+  {
+    href: '/examples/recommendations.html',
+    label: 'Recommendations',
+  },
   {href: '/examples/horizontal-facets.html', label: 'Horizontal Facets'},
   {
     href: '/accessibility/commerce-full.html',