+++ packages/atomic/src/utils/props-utils.ts
@@ -1,3 +1,4 @@
+import {isArray} from '@coveo/bueno';
 import {ComponentInterface, getElement} from '@stencil/core';
 import {camelToKebab, kebabToCamel} from './utils';
 
@@ -5,6 +6,13 @@ interface MapPropOptions {
   attributePrefix?: string;
   splitValues?: boolean;
 }
+
+interface ArrayPropOptions {
+  // @deprecated
+  // TODO v2: remove deprecation warning and change to a strict error only
+  deprecationWarning: boolean;
+}
+
 export function MapProp(opts?: MapPropOptions) {
   return (component: ComponentInterface, variableName: string) => {
     const {componentWillLoad} = component;
@@ -30,6 +38,67 @@ export function MapProp(opts?: MapPropOptions) {
   };
 }
 
+export function ArrayProp(opts: ArrayPropOptions) {
+  const logWithDeprecation = (msg: string, ...other: unknown[]) =>
+    opts.deprecationWarning
+      ? console.warn(
+          `${msg} This will be enforced in the next major version`,
+          other
+        )
+      : console.error(msg, other);
+
+  return (component: ComponentInterface, variableName: string) => {
+    const {componentWillLoad} = component;
+    const attributeWithBackets = camelToKebab(variableName);
+
+    if (!componentWillLoad) {
+      console.error(
+        'The "componentWillLoad" lifecycle method has to be defined for the ArrayProp decorator to work.'
+      );
+      return;
+    }
+
+    component.componentWillLoad = function () {
+      const attr =
+        getElement(this).attributes.getNamedItem(attributeWithBackets);
+      if (!attr) {
+        componentWillLoad.call(this);
+        return;
+      }
+
+      try {
+        const valueAsArray = JSON.parse(attr.value);
+        if (isArray(valueAsArray)) {
+          this[variableName] = valueAsArray;
+        } else {
+          logWithDeprecation(
+            `Property ${attributeWithBackets} should be an array`,
+            getElement(this)
+          );
+        }
+      } catch (e) {
+        logWithDeprecation(
+          `Error while parsing attribute ${attributeWithBackets} as array`,
+          e
+        );
+      }
+
+      componentWillLoad.call(this);
+    };
+  };
+}
+
+function splitAttributeValueOnCommas(attributeValue: string) {
+  const splitButIgnoreEscapeSymbolsExpression = /(?:\\.|[^,])+/g;
+  const [...valuesWithEscapeSymbols] =
+    attributeValue.matchAll(splitButIgnoreEscapeSymbolsExpression) ?? [];
+
+  const removeEscapeSymbolsExpression = /\\(.)/g;
+  return valuesWithEscapeSymbols.map(([valuesWithEscapeSymbols]) =>
+    valuesWithEscapeSymbols.replace(removeEscapeSymbolsExpression, '$1')
+  );
+}
+
 export function mapAttributesToProp(
   prefix: string,
   mapVariable: Record<string, string | string[]>,
@@ -47,7 +116,9 @@ function stringMapToStringArrayMap(map: Record<string, string>) {
   return Object.entries(map).reduce(
     (acc, [key, value]) => ({
       ...acc,
-      [key]: value.split(',').map((subValue) => subValue.trim()),
+      [key]: splitAttributeValueOnCommas(value).map((subValue) =>
+        subValue.trim()
+      ),
     }),
     {}
   );