+++ packages/headless/src/controllers/insight/tab/headless-insight-tab.ts
@@ -7,7 +7,7 @@ import {
   TabOptions,
 } from '../../core/tab/headless-core-tab';
 import {InsightEngine} from '../../../app/insight-engine/insight-engine';
-import {logInsightInterfaceChange} from '../../../features/analytics/analytics-actions';
+import {logInsightInterfaceChange} from '../../../features/analytics/insight-analytics-actions';
 import {executeSearch} from '../../../features/insight-search/insight-search-actions';
 
 export type {Tab, TabProps, TabState, TabInitialState, TabOptions};